requirements.txt
faiss-cpu==1.7.4
langchain==0.2.12
langchain-openai==0.1.21 
langchain-community==0.2.11
matplotlib==3.6.0
numpy==1.24.4
#openai==1.9.0
openpyxl==3.1.0
pandas==2.2.2
streamlit==1.37.1
tiktoken==0.7.0
#tokenizers==0.12.1
#transformers==4.30.2